requirements.txt
pytest==7.4.3
boto3==1.34.0
aws-lambda-powertools==2.30.1
pyyaml==6.0
